Mudado para Long o tipo do parâmetro de deletaEvento()

+++ app/src/main/java/com/infact/nightour/controller/EventosController.java
@@ -95,7 +95,7 @@ public class EventosController {
         db.close();
     }
 
-    public void deletaEvento(int id) {
+    public void deletaEvento(Long id) {
         String where = Evento.BD_ID + " = " + id;
 
         db = banco.getWritableDatabase();
